refactor: 원시값 포장하기
rename Bowl to Pins
